requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.1
djangorestframework==3.14.0
Faker==22.5.1
gunicorn==21.2.0
idna==3.6
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
